Fixed convertion error

+++ src/settings.h
@@ -2,6 +2,6 @@
 #define SETTINGS_H
 
 const int WIDTH = 1280, HEIGHT = 720;
-const float screenRatio = WIDTH / HEIGHT;
+const float screenRatio = (float) WIDTH / HEIGHT;
 
 #endif
